+++ java/code/src/com/redhat/rhn/taskomatic/task/CobblerSyncTask.java
@@ -65,9 +65,9 @@ public class CobblerSyncTask extends SingleThreadedTestableTask {
         XMLRPCInvoker invoker = (XMLRPCInvoker)  
             MethodUtil.getClassFromConfig(CobblerXMLRPCHelper.class.getName());
         
-        Integer mtime = null;
+        Double mtime = null;
         try {
-            mtime = (Integer) invoker.invokeMethod("last_modified_time", new ArrayList());
+            mtime = (Double) invoker.invokeMethod("last_modified_time", new ArrayList());
         }
         catch (XmlRpcFault e) {
             log.error("Error calling cobbler.", e);
